moves data entry e2e tests to separate folder
