+++ frontend/e2e-tests/tests/data-entry/data-entry.valid.model.e2e.ts
@@ -9,8 +9,11 @@ import { createMachine } from "xstate";
 
 import { VotersCounts, VotesCounts } from "@/types/generated/openapi";
 
-import { test } from "../fixtures";
-import { getStatesAndEventsFromMachineDefinition, getStatesAndEventsFromTest } from "../helpers-utils/xstate-helpers";
+import { test } from "../../fixtures";
+import {
+  getStatesAndEventsFromMachineDefinition,
+  getStatesAndEventsFromTest,
+} from "../../helpers-utils/xstate-helpers";
 
 /*
 The names of the states in the machine keep track of two states:
