+++ frontend/e2e-tests/tests/data-entry/resume-data-entry.e2e.ts
@@ -8,9 +8,9 @@ import { VotersAndVotesPage } from "e2e-tests/page-objects/data_entry/VotersAndV
 
 import { PollingStation, VotersCounts, VotesCounts } from "@/types/generated/openapi";
 
-import { test } from "../fixtures";
-import { loginAs } from "../setup";
-import { emptyDataEntryResponse } from "../test-data/request-response-templates";
+import { test } from "../../fixtures";
+import { loginAs } from "../../setup";
+import { emptyDataEntryResponse } from "../../test-data/request-response-templates";
 
 test.use({
   storageState: "e2e-tests/state/typist.json",
